backend/microservice-TTS/requirements.txt
fastapi
uvicorn
httpx
aioboto3
pydantic
google-genai